https://issues.apache.org/activemq/browse/AMQCPP-103
Flushing out decaf build issues on windows.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@648805 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/cmsutil/CmsTemplate.cpp
@@ -617,7 +617,7 @@ throw (cms::CMSException) {
             return consumer->receive();
         }
         default: {
-            return consumer->receive(receiveTime);
+            return consumer->receive((int)receiveTime);
         }
         }
